[dsub] Bump version, refactor jobs_controller (#142)
