+++ servers/dsub/jobs/controllers/utils/labels.py
@@ -1,0 +1,19 @@
+def dsub_to_api(job):
+    """Extracts labels from a job, if present.
+
+        Args:
+            job: A dict with dsub job metadata
+
+        Returns:
+            dict: Labels key value pairs with dsub-specific information
+    """
+    # Put any dsub specific information into the labels. These fields are
+    # candidates for the common jobs API
+    labels = job['labels'].copy() if job['labels'] else {}
+    if 'status-detail' in job:
+        labels['status-detail'] = job['status-detail']
+    if 'last-update' in job:
+        labels['last-update'] = job['last-update']
+    if 'user-id' in job:
+        labels['user-id'] = job['user-id']
+    return labels
