+++ servers/dsub/jobs/test/base_test_cases.py
@@ -10,7 +10,7 @@ import string
 import time
 
 from jobs.common import execute_redirect_stdout
-from jobs.controllers.job_statuses import ApiStatus
+from jobs.controllers.utils.job_statuses import ApiStatus
 from jobs.encoder import JSONEncoder
 from jobs.models.job_metadata_response import JobMetadataResponse
 from jobs.models.query_jobs_response import QueryJobsResponse
@@ -107,32 +107,33 @@ class BaseTestCases:
             resources = job_util.JobResources(
                 image=DOCKER_IMAGE, logging=logging, zones=['us-central1*'])
 
-            env_data = [param_util.EnvParam(k, v) for (k, v) in envs.items()]
-            label_data = [
-                param_util.LabelParam(k, v) for (k, v) in labels.items()
-            ]
+            env_data = {param_util.EnvParam(k, v) for (k, v) in envs.items()}
+            label_data = {
+                param_util.LabelParam(k, v)
+                for (k, v) in labels.items()
+            }
 
             # This is mostly an extraction dsubs argument parsing here:
             # https://github.com/googlegenomics/dsub/blob/master/dsub/lib/param_util.py#L720
             # Reworked it to handle dictionaries rather than a list of items
             # of the form 'key=val'
             input_file_param_util = param_util.InputFileParamUtil('input')
-            input_data = []
+            input_data = set()
             for (recursive, items) in ((False, inputs.items()),
                                        (True, inputs_recursive.items())):
                 for (name, value) in items:
                     name = input_file_param_util.get_variable_name(name)
-                    input_data.append(
+                    input_data.add(
                         input_file_param_util.make_param(
                             name, value, recursive))
 
             output_file_param_util = param_util.OutputFileParamUtil('output')
-            output_data = []
+            output_data = set()
             for (recursive, items) in ((False, outputs.items()),
                                        (True, outputs_recursive.items())):
                 for (name, value) in items:
                     name = output_file_param_util.get_variable_name(name)
-                    output_data.append(
+                    output_data.add(
                         output_file_param_util.make_param(
                             name, value, recursive))
 
@@ -274,6 +275,14 @@ class BaseTestCases:
                 [succeeded_job])
             self.assert_query_matches(
                 QueryJobsRequest(statuses=[ApiStatus.RUNNING]), [running_job])
+            self.assert_query_matches(
+                QueryJobsRequest(
+                    statuses=[ApiStatus.RUNNING, ApiStatus.SUCCEEDED]),
+                [succeeded_job, running_job])
+            self.assert_query_matches(
+                QueryJobsRequest(
+                    statuses=[ApiStatus.SUCCEEDED, ApiStatus.RUNNING]),
+                [succeeded_job, running_job])
 
         def test_query_jobs_by_label(self):
             labels = {
