+++ servers/dsub/requirements-to-freeze.txt
@@ -6,5 +6,5 @@ google-compute-engine==2.7.2
 google-api-python-client == 1.5.5
 gunicorn == 19.7.1
 oauth2client == 3.0.0
-dsub == 0.1.3
+dsub == 0.1.4
 flask-cors == 3.0.3
