+++ servers/dsub/requirements.txt
@@ -4,7 +4,7 @@ chardet==3.0.4
 click==6.7
 clickclick==1.2.2
 connexion==1.1.13
-dsub==0.1.3
+dsub==0.1.4
 Flask==0.12.2
 Flask-Cors==3.0.3
 functools32==3.2.3-2
